Basic Reminder Creation and Most of the Timer Functionality With a Dialog Box
I was able to get basic reminders created and displayed. I was also able to get the timer to work, but I keep getting an exception when I try to create a dialog box (It says that you can't create Java FX elements unless it's from the main Java FX thread).