Fixed the dialog box bug, got a main_loop() working, and added a todo list.
I managed to sort out the bug where the dialog box couldn't be created. The solution was as unintuitive as it possibly could've been. Anyway, as a result I was able to create a function called main_loop() that would get run once every 0.75 seconds, which was nice for checking the timer.  I also created a todo list to outline what sorts of things are left to be done.