WindowsBuild
I added a way of building an executable which'll run the application. I also create a batch file called create_distributable_program.bat. This will create a folder called MedReminder and put all of the necessary files to run the program in it, including the executable. This is just so it's easy to run.